Restructure and document pure tree implementation
